+++ dffml/operation/model.py
@@ -5,6 +5,8 @@ from ..base import config
 from ..model import Model
 from ..df.types import Definition
 from ..df.base import op
+from ..source.source import Sources
+from ..source.memory import MemorySource, MemorySourceConfig
 
 
 @config
@@ -95,8 +97,13 @@ async def model_predict(self, features: Dict[str, Any]) -> Dict[str, Any]:
     {'model_predictions': {'Salary': {'confidence': 1.0, 'value': 50.0}}}
     """
 
-    async def records():
-        yield Record("", data={"features": features})
-
-    async for record in self.mctx.predict(records()):
-        return {"prediction": record.predictions()}
+    async with Sources(
+        MemorySource(
+            MemorySourceConfig(
+                records=[Record("", data={"features": features})]
+            )
+        )
+    ) as source:
+        async with source() as sctx:
+            async for record in self.mctx.predict(sctx):
+                return {"prediction": record.predictions()}