+++ model/transformers/tests/test_ner_model.py
@@ -84,7 +84,7 @@ class TestNERModel(AsyncTestCase):
         async with self.predict_sources as sources, self.model as model:
             target_name = model.config.predict.name
             async with sources() as sctx, model() as mctx:
-                async for record in mctx.predict(sctx.records()):
+                async for record in mctx.predict(sctx):
                     prediction = record.prediction(target_name).value
                     for d in prediction:
                         self.assertIn(